Replace kube object menu registry with reactive solution
