+++ src/extensions/registries/index.ts
@@ -8,7 +8,6 @@
 export * from "./page-registry";
 export * from "./page-menu-registry";
 export * from "./kube-object-detail-registry";
-export * from "./kube-object-menu-registry";
 export * from "./kube-object-status-registry";
 export * from "./entity-setting-registry";
 export * from "./catalog-entity-detail-registry";
