+++ src/renderer/components/+workloads-cronjobs/cronjobs.tsx
@@ -8,20 +8,15 @@ import "./cronjobs.scss";
 import React from "react";
 import { observer } from "mobx-react";
 import type { RouteComponentProps } from "react-router";
-import { CronJob, cronJobApi } from "../../../common/k8s-api/endpoints/cron-job.api";
-import { MenuItem } from "../menu";
-import { Icon } from "../icon";
+import type { CronJob } from "../../../common/k8s-api/endpoints/cron-job.api";
 import { cronJobStore } from "./cronjob.store";
 import { jobStore } from "../+workloads-jobs/job.store";
 import { eventStore } from "../+events/event.store";
-import type { KubeObjectMenuProps } from "../kube-object-menu";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
-import { CronJobTriggerDialog } from "./cronjob-trigger-dialog";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
-import { ConfirmDialog } from "../confirm-dialog/confirm-dialog";
-import { Notifications } from "../notifications/notifications";
 import type { CronJobsRouteParams } from "../../../common/routes";
 import moment from "moment";
+import { CronJobMenu } from "./cron-job-menu";
 
 enum columnId {
   name = "name",
@@ -87,53 +82,3 @@ export class CronJobs extends React.Component<Props> {
   }
 }
 
-export function CronJobMenu(props: KubeObjectMenuProps<CronJob>) {
-  const { object, toolbar } = props;
-
-  return (
-    <>
-      <MenuItem onClick={() => CronJobTriggerDialog.open(object)}>
-        <Icon material="play_circle_filled" tooltip="Trigger" interactive={toolbar}/>
-        <span className="title">Trigger</span>
-      </MenuItem>
-
-      {object.isSuspend() ?
-        <MenuItem onClick={() => ConfirmDialog.open({
-          ok: async () => {
-            try {
-              await cronJobApi.resume({ namespace: object.getNs(), name: object.getName() });
-            } catch (err) {
-              Notifications.error(err);
-            }
-          },
-          labelOk: `Resume`,
-          message: (
-            <p>
-              Resume CronJob <b>{object.getName()}</b>?
-            </p>),
-        })}>
-          <Icon material="play_circle_outline" tooltip="Resume" interactive={toolbar}/>
-          <span className="title">Resume</span>
-        </MenuItem>
-
-        : <MenuItem onClick={() => ConfirmDialog.open({
-          ok: async () => {
-            try {
-              await cronJobApi.suspend({ namespace: object.getNs(), name: object.getName() });
-            } catch (err) {
-              Notifications.error(err);
-            }
-          },
-          labelOk: `Suspend`,
-          message: (
-            <p>
-              Suspend CronJob <b>{object.getName()}</b>?
-            </p>),
-        })}>
-          <Icon material="pause_circle_filled" tooltip="Suspend" interactive={toolbar}/>
-          <span className="title">Suspend</span>
-        </MenuItem>
-      }
-    </>
-  );
-}
