+++ src/renderer/components/+workloads-deployments/deployments.tsx
@@ -8,12 +8,7 @@ import "./deployments.scss";
 import React from "react";
 import { observer } from "mobx-react";
 import type { RouteComponentProps } from "react-router";
-import { Deployment, deploymentApi } from "../../../common/k8s-api/endpoints";
-import type { KubeObjectMenuProps } from "../kube-object-menu";
-import { MenuItem } from "../menu";
-import { Icon } from "../icon";
-import { DeploymentScaleDialog } from "./deployment-scale-dialog";
-import { ConfirmDialog } from "../confirm-dialog";
+import type { Deployment } from "../../../common/k8s-api/endpoints";
 import { deploymentStore } from "./deployments.store";
 import { eventStore } from "../+events/event.store";
 import { KubeObjectListLayout } from "../kube-object-list-layout";
@@ -21,8 +16,8 @@ import { cssNames } from "../../utils";
 import kebabCase from "lodash/kebabCase";
 import orderBy from "lodash/orderBy";
 import { KubeObjectStatusIcon } from "../kube-object-status-icon";
-import { Notifications } from "../notifications";
 import type { DeploymentsRouteParams } from "../../../common/routes";
+import { DeploymentMenu } from "./deployment-menu";
 
 enum columnId {
   name = "name",
@@ -94,39 +89,4 @@ export class Deployments extends React.Component<Props> {
       />
     );
   }
-}
-
-export function DeploymentMenu(props: KubeObjectMenuProps<Deployment>) {
-  const { object, toolbar } = props;
-
-  return (
-    <>
-      <MenuItem onClick={() => DeploymentScaleDialog.open(object)}>
-        <Icon material="open_with" tooltip="Scale" interactive={toolbar}/>
-        <span className="title">Scale</span>
-      </MenuItem>
-      <MenuItem onClick={() => ConfirmDialog.open({
-        ok: async () =>
-        {
-          try {
-            await deploymentApi.restart({
-              namespace: object.getNs(),
-              name: object.getName(),
-            });
-          } catch (err) {
-            Notifications.error(err);
-          }
-        },
-        labelOk: `Restart`,
-        message: (
-          <p>
-            Are you sure you want to restart deployment <b>{object.getName()}</b>?
-          </p>
-        ),
-      })}>
-        <Icon material="autorenew" tooltip="Restart" interactive={toolbar}/>
-        <span className="title">Restart</span>
-      </MenuItem>
-    </>
-  );
 }
