+++ src/renderer/components/kube-object-menu/dependencies/kube-object-menu-items/get-kube-object-menu-items.ts
@@ -2,21 +2,36 @@
  * Copyright (c) OpenLens Authors. All rights reserved.
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
-import type { KubeObjectMenuRegistry } from "../../../../../extensions/registries";
+import { conforms, includes, eq } from "lodash/fp";
 import type { KubeObject } from "../../../../../common/k8s-api/kube-object";
+import type { LensRendererExtension } from "../../../../../extensions/lens-renderer-extension";
+import { staticKubeObjectMenuItems as staticMenuItems } from "./static-kube-object-menu-items";
+
+interface Dependencies {
+  extensions: LensRendererExtension[];
+  kubeObject: KubeObject;
+}
 
 export const getKubeObjectMenuItems = ({
-  kubeObjectMenuRegistry,
+  extensions,
   kubeObject,
-}: {
-  kubeObjectMenuRegistry: KubeObjectMenuRegistry;
-  kubeObject: KubeObject;
-}) => {
+}: Dependencies) => {
   if (!kubeObject) {
     return [];
   }
 
-  return kubeObjectMenuRegistry
-    .getItemsForKind(kubeObject.kind, kubeObject.apiVersion)
+  const extensionMenuItems = extensions.flatMap(
+    (extension) => extension.kubeObjectMenuItems,
+  );
+
+  return [...staticMenuItems, ...extensionMenuItems]
+    .filter(
+      conforms({
+        kind: eq(kubeObject.kind),
+        apiVersions: includes(kubeObject.apiVersion),
+      }),
+    )
+
     .map((item) => item.components.MenuItem);
 };
+
