+++ src/renderer/components/kube-object-menu/dependencies/kube-object-menu-items/kube-object-menu-items.injectable.ts
@@ -3,15 +3,15 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable, lifecycleEnum } from "@ogre-tools/injectable";
-import kubeObjectMenuRegistryInjectable from "./kube-object-menu-registry.injectable";
 
 import { getKubeObjectMenuItems } from "./get-kube-object-menu-items";
 import type { KubeObject } from "../../../../../common/k8s-api/kube-object";
+import rendererExtensionsInjectable from "../../../../../extensions/renderer-extensions.injectable";
 
 const kubeObjectMenuItemsInjectable = getInjectable({
   instantiate: (di, { kubeObject }: { kubeObject: KubeObject }) =>
     getKubeObjectMenuItems({
-      kubeObjectMenuRegistry: di.inject(kubeObjectMenuRegistryInjectable),
+      extensions: di.inject(rendererExtensionsInjectable).get(),
       kubeObject,
     }),
