--- src/renderer/components/kube-object-menu/dependencies/kube-object-menu-items/kube-object-menu-registry.injectable.ts
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) OpenLens Authors. All rights reserved.
- * Licensed under MIT License. See LICENSE in root directory for more information.
- */
-import { KubeObjectMenuRegistry } from "../../../../../extensions/registries";
-import { getInjectable, lifecycleEnum } from "@ogre-tools/injectable";
-
-const kubeObjectMenuRegistryInjectable = getInjectable({
-  instantiate: () => KubeObjectMenuRegistry.getInstance(),
-  lifecycle: lifecycleEnum.singleton,
-});
-
-export default kubeObjectMenuRegistryInjectable;
