+++ src/renderer/initializers/index.ts
@@ -8,7 +8,6 @@ export * from "./catalog";
 export * from "./entity-settings-registry";
 export * from "./ipc";
 export * from "./kube-object-detail-registry";
-export * from "./kube-object-menu-registry";
 export * from "./registries";
 export * from "./workloads-overview-detail-registry";
 export * from "./catalog-category-registry";
