+++ src/renderer/initializers/registries.ts
@@ -12,7 +12,6 @@ export function initRegistries() {
   registries.EntitySettingRegistry.createInstance();
   registries.GlobalPageRegistry.createInstance();
   registries.KubeObjectDetailRegistry.createInstance();
-  registries.KubeObjectMenuRegistry.createInstance();
   registries.KubeObjectStatusRegistry.createInstance();
   registries.WorkloadsOverviewDetailRegistry.createInstance();
 }
